converting js to ts
