+++ transaction/splits/unwrap.ts
@@ -1,25 +1,24 @@
-const keys = require("../../utilities/keys");
-const broadcast = require("../../utilities/broadcastTx");
-const config = require("../../config.json");
-const request = require("request");
-const Promise = require("promise");
-const persistenceClass = require("../../utilities/persistenceJS");
+import * as config from "../../config.json";
+import { request } from "request";
+import { Persistence } from "../../utilities/persistenceJS";
+import { broadcastTx } from "../../utilities/broadcastTx";
+import { getWallet } from "../../utilities/keys";
 
-class unwrapsplits extends persistenceClass {
-  async unwrap(
-    address,
-    chain_id,
-    mnemonic,
-    fromID,
-    ownableID,
-    split,
-    feesAmount,
-    feesToken,
-    gas,
-    mode,
-    memo = ""
-  ) {
-    const wallet = keys.getWallet(mnemonic);
+class unwrapsplits extends Persistence {
+  unwrap = async (
+      address: string,
+      chain_id: string,
+      mnemonic: string,
+      fromID: string,
+      ownableID: string,
+      split: any,
+      feesAmount: any,
+      feesToken: any,
+      gas: any,
+      mode: any,
+      memo: string
+  ): Promise<any> => {
+    const wallet = getWallet(mnemonic, "");
     let path = this.path;
 
     let options = {
@@ -52,7 +51,7 @@ class unwrapsplits extends persistenceClass {
         }
         let result = JSON.parse(response.body);
         resolve(
-          broadcast.broadcastTx(path, wallet, result.value, chain_id, mode)
+          broadcastTx(path, wallet, result.value, chain_id, mode)
         );
       });
     }).catch(function (error) {
